fix bug of quatd * int; combine hpp and cpp; fix << overload error in win system; modify include in test file;
