+++ modules/core/src/precomp.hpp
@@ -87,7 +87,6 @@
 #include "opencv2/core/neon_utils.hpp"
 #include "opencv2/core/vsx_utils.hpp"
 #include "hal_replacement.hpp"
-#include "opencv2/core/quaternion.hpp"
 
 #define GET_OPTIMIZED(func) (func)
